+++ particular/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MassiveAffected` no longer implements `Storage`. `MassiveAffectedInternal` and `MassiveAffectedArray` newtypes are used instead.
 - `BarnesHutTree` trait renamed to `BarnesHutAcceleration`.
 - `IntoPointMass::point_mass` takes ownership of `self`.
+- `Storage::store` and `ParticleStorage::store_particles` parameter `I` bound to `IntoIterator<Item = P>`.
 
 ### Removed